apigee-operator/requirements.txt
kopf==1.37.1
google-auth==2.28.2
xmltodict==0.13.0